Change from integer to double division to keep decimal place (so 1.3 rather than 1.0)
git-svn-id: https://svn.code.sf.net/p/autowikibrowser/code@3520 5227b50d-9349-4574-b055-4b0a7525e11c

+++ AWB/AWB/Main.cs
@@ -1521,7 +1521,7 @@ window.scrollTo(0, diffTopY);
 
             if (OverallTypoStats.Saves > 0)
                 lblTypoRatio.Text = string.Format("{0:F1}",
-                    (total - selfMatches - OverallTypoStats.FalsePositives) / OverallTypoStats.Saves);
+                    ((double)(total - selfMatches - OverallTypoStats.FalsePositives) / (double)OverallTypoStats.Saves));
             else
                 lblTypoRatio.Text = "0";
         }
